+++ src/parser/block/paragraph/hashtag.ts
@@ -1,18 +1,17 @@
 ﻿import { ParagraphParser } from '../../block';
 import { combine, some } from '../../../combinator';
 import { unescsource } from '../../source/unescapable';
+import { index } from '../../inline';
 import { squash } from '../../squash';
 import { match } from '../../source/validation';
 import { html } from 'typed-dom';
 
 const syntax = /^#\S+/;
-const escape = /^\S#+/;
 const closer = /^\s/;
 
 export const hashtag: ParagraphParser.HashtagParser = source => {
-  const [flag = undefined] = source.match(escape) || [];
-  if (flag) return [[document.createTextNode(flag)], source.slice(flag.length)];
   if (!match(source, '#', syntax)) return;
+  if (index(source)) return;
   const line = source.split('\n', 1)[0];
   const [ts = [], rest = undefined] = some(combine<ParagraphParser.HashtagParser>([unescsource]), closer)(line) || [];
   if (rest === undefined) return;